--- Elements/src/DoorOpeningSide.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Elements
-{
-    public enum DoorOpeningSide
-    {
-        [System.Runtime.Serialization.EnumMember(Value = @"Undefined")]
-        Undefined,
-        [System.Runtime.Serialization.EnumMember(Value = @"Left Hand")]
-        LeftHand,
-        [System.Runtime.Serialization.EnumMember(Value = @"Right Hand")]
-        RightHand,
-        [System.Runtime.Serialization.EnumMember(Value = @"Double Door")]
-        DoubleDoor
-    }
-}
