--- Elements/src/DoorOpeningType.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Elements
-{
-    public enum DoorOpeningType
-    {
-        [System.Runtime.Serialization.EnumMember(Value = @"Undefined")]
-        Undefined,
-        [System.Runtime.Serialization.EnumMember(Value = @"Single Swing")]
-        SingleSwing,
-        [System.Runtime.Serialization.EnumMember(Value = @"Double Swing")]
-        DoubleSwing
-    }
-}
