+++ Elements/test/DoorTest.cs
@@ -18,7 +18,7 @@ namespace Elements
 
             var line = new Line(new Vector3(0, 0, 0), new Vector3(10, 10, 0));
             var wall = new StandardWall(line, 0.1, 3.0);
-            var door = new Door(wall.CenterLine, 0.5, 2.0, 2.0, Door.DOOR_DEFAULT_THICKNESS, DoorOpeningSide.LeftHand, DoorOpeningType.SingleSwing);
+            var door = new Door(wall.CenterLine, 0.5, 2.0, 2.0, Door.DOOR_THICKNESS, DoorOpeningSide.LeftHand, DoorOpeningType.SingleSwing);
             wall.AddDoorOpening(door);
 
             this.Model.AddElement(wall);
